refactor(api): rename acp notifications and import result to types.ts
Aligns with the project convention that every contract module exposes
its schemas and types via a types.ts file (see brand/types.ts,
auth/types.ts, files/types.ts, skills/types.ts).
